+++ templates/landingpage/design.js
@@ -49,6 +49,8 @@ module.exports = cx.design
         require('@bsi-cx/design-standard-library-web/content-elements/base/video'),
         require('@bsi-cx/design-standard-library-web/content-elements/base/social-media-follow'),
         require('@bsi-cx/design-standard-library-web/content-elements/base/social-media-share'),
+        require('@bsi-cx/design-standard-library-web/content-elements/advanced/webcam-image-upload'),
+        require('@bsi-cx/design-standard-library-web/content-elements/base/slot-finder'),
         require('@bsi-cx/design-standard-library-web/content-elements/base/chart'),
         require('@bsi-cx/design-standard-library-web/content-elements/base/html')),
     cx.contentElementGroup